requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.20.0
albumentations>=1.3.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep learning frameworks and utilities
timm>=0.9.0
transformers>=4.30.0
segmentation-models-pytorch>=0.3.0

# Data handling
h5py>=3.8.0
zarr>=2.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Configuration and utilities
PyYAML>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0
rich>=13.4.0
tqdm>=4.65.0

# Utilities
click>=8.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0
pre-commit>=3.3.0

# Medical image processing (optional)
SimpleITK>=2.2.0
pydicom>=2.4.0

# GPU acceleration (optional)
cupy-cuda11x>=12.0.0  # Adjust based on CUDA version

# Additional ML utilities
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Graph processing
networkx>=3.1.0

# Compression
hdf5storage>=0.1.18

# Data validation
jsonschema>=4.18.0

# Parallel processing
joblib>=1.3.0
multiprocessing-logging>=0.3.4

# Memory optimization
psutil>=5.9.0
memory-profiler>=0.60.0

# Pathlib extensions
pathlib2>=2.3.7